apps/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
pgvector==0.2.3

# AI and Voice Processing
openai==1.3.7
anthropic==0.7.7
langchain==0.0.350
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-anthropic==0.0.1
pinecone-client==2.2.4
weaviate-client==3.25.3

# Speech Recognition and Processing
speechrecognition==3.10.0
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
whisper==1.1.10

# TTS
elevenlabs==0.2.26

# Calendar Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
icalendar==5.0.7
dateutil==2.8.2
msal==1.24.1

# Real-time Communication
websockets==12.0
socketio==5.10.0
python-socketio==5.10.0

# Background Tasks
celery==5.3.4
flower==2.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Data Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic-extra-types==2.1.0
email-validator==2.1.0

# Security
cryptography==41.0.8
bcrypt==4.1.2

# File Processing
python-magic==0.4.27
Pillow==10.1.0

# Email and Notifications
fastapi-mail==1.4.1
twilio==8.10.0

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Environment and Configuration
python-decouple==3.8
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0
